+++ common/src/main/kotlin/com/metacontent/cobblenav/spawndata/collector/block/AreaTypeBlockCollector.kt
@@ -4,14 +4,15 @@ import com.cobblemon.mod.common.api.spawning.condition.AreaTypeSpawningCondition
 import com.cobblemon.mod.common.api.spawning.context.AreaSpawningContext
 import com.metacontent.cobblenav.spawndata.collector.BlockConditionCollector
 import com.metacontent.cobblenav.spawndata.collector.ConfigureableCollector
+import com.metacontent.cobblenav.util.ModDependency
 import net.minecraft.core.registries.BuiltInRegistries
 import net.minecraft.resources.ResourceLocation
 
 class AreaTypeBlockCollector : BlockConditionCollector<AreaTypeSpawningCondition<*>>, ConfigureableCollector {
     override val configName = "area_type_block"
     override val conditionClass = AreaTypeSpawningCondition::class.java
-    override var neededInstalledMods: List<String> = emptyList()
-    override var neededUninstalledMods: List<String> = emptyList()
+    override var neededInstalledMods: List<ModDependency> = emptyList()
+    override var neededUninstalledMods: List<ModDependency> = emptyList()
 
     override fun collect(condition: AreaTypeSpawningCondition<*>, contexts: List<AreaSpawningContext>): Set<ResourceLocation> {
         val blocks = mutableSetOf<ResourceLocation>()