+++ common/src/main/kotlin/com/metacontent/cobblenav/spawndata/collector/special/mythsandlegends/PokemonCollector.kt
@@ -5,14 +5,17 @@ import com.cobblemon.mod.common.api.spawning.context.SpawningContext
 import com.github.d0ctorleon.mythsandlegends.cobblemon.spawning.condition.pokemon.PokemonCondition
 import com.metacontent.cobblenav.api.platform.SpawnDataContext
 import com.metacontent.cobblenav.spawndata.collector.ConditionCollector
+import com.metacontent.cobblenav.spawndata.collector.ConfigureableCollector
+import com.metacontent.cobblenav.util.ModDependency
 import net.minecraft.network.chat.Component
 import net.minecraft.network.chat.MutableComponent
 import net.minecraft.server.level.ServerPlayer
 
-class PokemonCollector : ConditionCollector<SpawningCondition<*>> {
+class PokemonCollector : ConditionCollector<SpawningCondition<*>>, ConfigureableCollector {
+    override val configName = "pokemon"
     override val conditionClass = SpawningCondition::class.java
-    override var neededInstalledMods: List<String> = listOf("mythsandlegends")
-    override var neededUninstalledMods: List<String> = emptyList()
+    override var neededInstalledMods: List<ModDependency> = listOf(ModDependency("mythsandlegends", "1.8.0"))
+    override var neededUninstalledMods: List<ModDependency> = emptyList()
 
     override fun collect(
         condition: SpawningCondition<*>,